backend/requirements.txt
langchain
openai
python-dotenv
pyaudio
SpeechRecognition
pyttsx3
asyncio
aiohttp
websockets
langchain-openai
elevenlabs
torch
opencv-python-headless
deepface
tf-keras
langchain-community
faster_whisper
numpy
fastapi
